section state management, add podcasts and talks
